Move compabitility shims to their own files and folder
