+++ Source/FakeItEasy.Net35/Compatibility/ConcurrentDictionary.cs
