+++ Source/FakeItEasy.Net35/Compatibility/Tuple.cs
@@ -1,0 +1,10 @@
+﻿namespace System
+{
+    internal static class Tuple
+    {
+        public static Tuple<T1, T2> Create<T1, T2>(T1 item1, T2 item2)
+        {
+            return new Tuple<T1, T2>(item1, item2);
+        }
+    }
+}
