+++ Source/FakeItEasy.Net35/Compatibility/Tuple.of.T1.T2.cs
@@ -1,0 +1,15 @@
+﻿namespace System
+{
+    internal class Tuple<T1, T2>
+    {
+        public Tuple(T1 item1, T2 item2)
+        {
+            this.Item1 = item1;
+            this.Item2 = item2;
+        }
+
+        public T1 Item1 { get; private set; }
+
+        public T2 Item2 { get; private set; }
+    }
+}
